class factory, actions reflection, minify
