--- src/@types/PolicyServer/auth.ts
@@ -1,1 +1,0 @@
-export type AuthType = 'waltid' | 'oauth' | 'ldap'
