--- src/@types/PolicyServer/containerTypes.ts
@@ -1,5 +1,0 @@
-const TYPES = {
-  IPolicyHandler: Symbol.for('IPolicyHandler')
-}
-
-export { TYPES }
