--- src/@types/PolicyServer/policyServerTypes.ts
@@ -1,3 +1,0 @@
-export { AuthType } from './auth'
-export { IPolicyHandler } from './policyHandler'
-export { PolicyActionType, PolicyRequestPayload, PolicyRequestResponse } from './request'
