+++ src/@types/policy.ts
@@ -1,0 +1,9 @@
+export type PolicyRequestPayload = {
+  action: string
+} & Record<PropertyKey, any>
+
+export type PolicyRequestResponse = {
+  success: boolean
+  message?: string
+  httpStatus: number
+}
