--- src/handlers/oauth/actions/initializeHandler.ts
@@ -1,23 +1,0 @@
-import {
-  AuthType,
-  IPolicyHandler,
-  PolicyActionType,
-  PolicyRequestPayload,
-  PolicyRequestResponse
-} from '../../../@types/PolicyServer/policyServerTypes'
-import { buildNotImplementedRequestMessage } from '../../../utils/validateRequests.js'
-
-export class OauthInitializeHandler implements IPolicyHandler {
-  supportAuthType(authType: AuthType): boolean {
-    return authType === 'oauth'
-  }
-
-  supportActionType(policyActionType: PolicyActionType): boolean {
-    return policyActionType === 'initialize'
-  }
-
-  // eslint-disable-next-line require-await
-  async execute(requestPayload: PolicyRequestPayload): Promise<PolicyRequestResponse> {
-    return buildNotImplementedRequestMessage()
-  }
-}
