+++ src/handlers/oauthPolicyHandler.ts
@@ -1,0 +1,3 @@
+import { PolicyHandler } from '../policyHandler.js'
+
+export class OAuthPolicyHandler extends PolicyHandler {}
