--- src/handlers/waltid/actions/uploadHandler.ts
@@ -1,23 +1,0 @@
-import {
-  AuthType,
-  IPolicyHandler,
-  PolicyActionType,
-  PolicyRequestPayload,
-  PolicyRequestResponse
-} from '../../../@types/PolicyServer/policyServerTypes'
-import { buildNotImplementedRequestMessage } from '../../../utils/validateRequests.js'
-
-export class WaltIdUploadHandler implements IPolicyHandler {
-  supportAuthType(authType: AuthType): boolean {
-    return authType === 'waltid'
-  }
-
-  supportActionType(policyActionType: PolicyActionType): boolean {
-    return policyActionType === 'upload'
-  }
-
-  // eslint-disable-next-line require-await
-  async execute(requestPayload: PolicyRequestPayload): Promise<PolicyRequestResponse> {
-    return buildNotImplementedRequestMessage()
-  }
-}
