+++ src/services/policyRouterService.ts
@@ -1,18 +1,6 @@
-import {
-  AuthType,
-  IPolicyHandler,
-  PolicyActionType,
-  PolicyRequestPayload,
-  PolicyRequestResponse
-} from '../@types/PolicyServer/policyServerTypes'
 import { Request, Response } from 'express'
-import { container } from '../handlers/container.js'
-import { PolicyHandlerSelector } from '../handlers/policyHandlerSelector.js'
-import { TYPES } from '../@types/PolicyServer/containerTypes.js'
-import 'reflect-metadata'
-
-const handlers = container.getAll<IPolicyHandler>(TYPES.IPolicyHandler)
-const handlerSelector = new PolicyHandlerSelector(handlers)
+import { PolicyHandlerFactory } from '../policyHandlerFactory.js'
+import { PolicyRequestPayload, PolicyRequestResponse } from '../@types/policy'
 
 const authType = process.env.AUTH_TYPE || 'waltid'
 export async function handlePolicyRequest(
@@ -22,19 +10,13 @@ export async function handlePolicyRequest(
   console.log('Request body:', req.body)
   const { action, ...rest } = req.body
 
-  const handler = handlerSelector.selectHandler(
-    authType as AuthType,
-    action as PolicyActionType
-  )
-
-  if (!handler) {
-    const errorResponse: PolicyRequestResponse = {
+  const handler = PolicyHandlerFactory.createPolicyHandler(authType)
+  if (handler == null) {
+    res.status(404).json({
       success: false,
-      message: `No handler found for authType: "${authType}" and actionType: "${action}"`,
-      httpStatus: 400
-    }
-    res.status(errorResponse.httpStatus).json(errorResponse)
-    return
+      status: 404,
+      message: `Handler for auth type "${authType}" is not found.`
+    })
   }
 
   const payload: PolicyRequestPayload = { action, ...rest }
