--- src/utils/asyncHandler.ts
@@ -1,7 +1,0 @@
-import { Request, Response, NextFunction, RequestHandler } from 'express'
-
-export const asyncHandler =
-  (fn: RequestHandler): RequestHandler =>
-  (req: Request, res: Response, next: NextFunction) => {
-    Promise.resolve(fn(req, res, next)).catch(next)
-  }
